+++ src/boutros/matsu/compiler.clj
@@ -51,6 +51,8 @@
                     (every? keyword? x) (str (name a) \: (name b))
                     :else "cannot encode!"))
     (map? x) (sub-compiler x)
+    (instance? clojure.lang.LazySeq x) (for [part (interpose \space x)]
+                                         (encode part))
     :else (throw (Exception.
                    (format "Don't know how to encode %s in an SPARQL context" x)))))